+++ sdk/tables/azure-data-tables/tests/test_table_client.py
@@ -443,7 +443,7 @@ class StorageTableClientTest(TableTestCase):
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
             self.assertIn(
-                "TestApp/v1.0 azsdk-python-storage-table/{} Python/{} ({})".format(
+                "TestApp/v1.0 azsdk-python-data-tables/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
                     platform.platform()),
@@ -456,7 +456,7 @@ class StorageTableClientTest(TableTestCase):
         def callback(response):
             self.assertTrue('User-Agent' in response.http_request.headers)
             self.assertIn(
-                "TestApp/v2.0 TestApp/v1.0 azsdk-python-storage-table/{} Python/{} ({})".format(
+                "TestApp/v2.0 TestApp/v1.0 azsdk-python-data-tables/{} Python/{} ({})".format(
                     VERSION,
                     platform.python_version(),
                     platform.platform()),