+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -1,12 +1,19 @@
 # Release History
 
-## 1.4.0b8 (Unreleased)
+## 1.5.0b1 (Unreleased)
+### Added
+- Application authentication APIs from 1.4.0b7
+- `ManagedIdentityCredential` supports the latest version of App Service
+  ([#11346](https://github.com/Azure/azure-sdk-for-python/issues/11346))
+
+## 1.4.0 (2020-08-10)
+### Added
 - `DefaultAzureCredential` uses the value of environment variable
 `AZURE_CLIENT_ID` to configure a user-assigned managed identity.
 ([#10931](https://github.com/Azure/azure-sdk-for-python/issues/10931))
-- Renamed `VSCodeCredential` to `VisualStudioCodeCredential`
 
 ### Breaking Changes
+- Renamed `VSCodeCredential` to `VisualStudioCodeCredential`
 - Removed application authentication APIs added in 1.4.0 beta versions. These
   will be reintroduced in 1.5.0b1. Passing the keyword arguments below
   generally won't cause a runtime error, but the arguments have no effect.
@@ -249,7 +256,7 @@ its use in national clouds
 ### New features:
 - `AuthorizationCodeCredential` authenticates with a previously obtained
 authorization code. See Azure Active Directory's
-[authorization code documentation](https://docs.microsoft.com/en-us/azure/active-directory/develop/v2-oauth2-auth-code-flow)
+[authorization code documentation](https://docs.microsoft.com/azure/active-directory/develop/v2-oauth2-auth-code-flow)
 for more information about this authentication flow.
 - Multi-cloud support: client credentials accept the authority of an Azure Active
 Directory authentication endpoint as an `authority` keyword argument. Known